Added main file and assets
I added the main file and assets.
The program is currently broken but I have added a working version of
the code.
I am currently trying to rewrite the code to not use globals but I'm
makign a mess of it.
So, I have made a back-up of the old file where everything works and
will move changes over.