Partly fixed main file
I have partly fixed the main file, but it still uses globals and have
neatened up the old main file.
The main file now works and is beginning to not rely on globals.
The old main file has also been tidied up and the child attribute
removed from classes. This has been replaced by type() which just
returns that class that the object is an instance of.