Removed use of globals
I removed the use of globals but had to hack in a change for the
mouseClick class.
I may end up removing the class entirely if I can come up with a
better solution, but for now it's hacked in.
If I can find a way that doesn't use the class or doesn't run into the
current chicken and the egg problem that'd be great and what I switch to.