Split main file into separate files
I split the main file into seperate files for each class.
I also plan on adding a method to the infoBox that will cause it to
disappear after a period of time or if the player has clicked away.
However, I can't figure out how to work with unix time well and I don't
want to suffer with using date time.
The main problem that I'm facing is that I can't seem to have a variable
that is equal to time.time() without it constantly updating.
This makes it impossible to track how long since a certain event has
happened using this method.
I should google/ask Dad further on this and see if I can find a work
around.