Split the main file into seperate files
Each class is now it's own file and this should help make development
easier.
The main file now imports all of the other files and calls the main
function instead of having all of the classes in it.
Each file also contains one class and imports any dependancies that
class may have.