Added menu icon class and removal of infoboxes
I added a new menu icon class that creates an info box when clicked.
However I am hoping that in future it would have useful functionality
such a making troops, a minimap etc.

I also added a system that removes infoboxes as soon as they are
created.
This means that after clicking on the base and clicking away, the info
box will appear to disappear, allowing the player to focus on other
things.
I'm not sure if I'm happy with this implementation, but since a new
info box is created every frame it has no impact on the end result.

I have also changed the gameObject class to create a mask for every game
object since they will all need one.