Base now creates menu icon when clicked
The base now creates a menu icon when clicked.
When the base image is clicked it now creates a menu icon in the bottom
right hand corner.
This menu icon also creates a info box when clicked, but the message is
not modular.
This menu icon is also delted when you click of off the base.

However, the way that this is currently done means that you cannot click on
this custom menu icon, rendering it useless.
This is because the icon created this frame is deleted immediatley after
it is drawn, and is then another icon is created the next frame.
In order to fix this I need to find a more sophisticated way to decide
whether to delete a menu icon.
I could add a variable that tracks what the mouse has clicked on or a
variable for each object that says if they are clicked on and then based
on this only delete objects that are not clicked on.

I also changed the window.WIDTH variable to be the more calm
window.width.