Refactored and rm Old main
I refactored the code to now have an game.initGame method that creates
the needed instances of the mouseClick, base etc.

This means that these instances can reference each other through
self.game.otherInstance.

This comes in usefull for checking what mouseClick has last clicked on.
Currently mouseClick has a self.clicked variable that reference the last
object that it clicked on. So when checking whether to delete infoBoxes or
menuIcons, they can reference the self.game.mouseClick.clicked variable to
see if they were the last object to be clicked on. Based on this they will
then only delete themselves if they are not the last clicked on object.
This means that if you click on the base and then on an infoBox the infoBox
will not disappear.

I also added a debugging file but did not commit it as I didn't want to
have github tracking a massive debugging file. I may later add a blank
file like in Parkour-Shooter-Game to github to act as a placeholder, but
I have not currently done that.

Lastly, I removed the Old_main file as it was no longer needed.